+++ packages/wells/src/__tests__/integration/wells.int.spec.ts
@@ -2,7 +2,7 @@
 
 import { setupLoggedInClient } from '../testUtils';
 import WellsClient from 'wells/src/client/cogniteWellsClient';
-import { Well, WellItems } from 'wells/src/client/model/Well';
+import { Well, WellItems, WellsLimits } from 'wells/src/client/model/Well';
 import { Wellbore } from 'wells/src/client/model/Wellbore';
 import { LengthRange } from 'wells/src/client/model/LengthRange';
 import {
@@ -68,6 +68,17 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
     expect(well.waterDepth?.unit).toBe('meter');
     expect(well.waterDepth?.value).toBe(100.0);
     /* eslint-enable */
+  });
+
+  test('get wells limits', async() => {
+    expect(client).not.toBeUndefined();
+    const limits: WellsLimits = await client.wells.limits()
+    expect(limits.spudDate).not.toBeUndefined()
+    expect(limits.waterDepth).not.toBeUndefined()
+    expect(limits.spudDate.max).toBeInstanceOf(Date)
+    expect(limits.spudDate.min).toBeInstanceOf(Date)
+    expect(limits.waterDepth.max.value).toBeGreaterThanOrEqual(limits.waterDepth.min.value)
+    expect(limits.nptDuration.max!).toBeGreaterThanOrEqual(limits.nptDuration.min!)
   });
 
   test('get source assets for well', async () => {
@@ -140,8 +151,28 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
     });
   });
 
+  test('list with limit, then cursor and limit', async () => {
+    const wellsWithLimit: WellItems = await client.wells.list(undefined, 3);
+    expect(wellsWithLimit).not.toBeUndefined();
+    expect(wellsWithLimit.items.length).toBe(3);
+
+    const wellsWithCursor: WellItems = await client.wells.list(wellsWithLimit.nextCursor, 2);
+    expect(wellsWithCursor).not.toBeUndefined();
+    expect(wellsWithCursor.items.length).toBe(2);
+  });
+
+  test('filter with limit, then cursor and limit', async () => {
+    const filter: WellFilter = { };
+    const wellsWithLimit: WellItems = await client.wells.filter(filter, undefined, 3);
+    expect(wellsWithLimit).not.toBeUndefined();
+    expect(wellsWithLimit.items.length).toBe(3);
+
+    const wellsWithCursor: WellItems = await client.wells.filter(filter, wellsWithLimit.nextCursor, 2);
+    expect(wellsWithCursor).not.toBeUndefined();
+    expect(wellsWithCursor.items.length).toBe(2);
+  });
+
   test('use cursor to get more wells', async () => {
-    expect(client).not.toBeUndefined();
     const wells = await client.wells.list();
     expect(wells).not.toBeUndefined();
     const retrievedWells = wells.items.map(x => x.id);
@@ -154,7 +185,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - gets wells in wkt polygon', async () => {
-    expect(client).not.toBeUndefined();
     const testPolygon =
       'POLYGON ((0.0 0.0, 0.0 80.0, 80.0 80.0, 80.0 0.0, 0.0 0.0))';
     const filter: WellFilter = {
@@ -171,7 +201,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - fuzzy search on name string matching 1', async () => {
-    expect(client).not.toBeUndefined();
     const filter: WellFilter = { stringMatching: '24' };
     const wells = await client.wells.filter(filter);
 
@@ -184,7 +213,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - fuzzy search on name string matching 2', async () => {
-    expect(client).not.toBeUndefined();
     const filter: WellFilter = { stringMatching: '10' };
     const wells = await client.wells.filter(filter);
 
@@ -197,7 +225,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter wells slow', async () => {
-    expect(client).not.toBeUndefined();
     const filter: WellFilter = { stringMatching: '10' };
     const wells: Well[] = await client.wells.filterSlow(filter);
 
@@ -217,7 +244,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - gets wells with description filter', async () => {
-    expect(client).not.toBeUndefined();
     const filter: WellFilter = { stringMatching: 'WDL Asset' };
     const wells = await client.wells.filter(filter);
 
@@ -230,7 +256,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - gets wells in polygon with description', async () => {
-    expect(client).not.toBeUndefined();
     const testPolygon =
       'POLYGON ((0.0 0.0, 0.0 80.0, 80.0 80.0, 80.0 0.0, 0.0 0.0))';
     const filter: WellFilter = {
@@ -249,7 +274,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - gets wells in polygon with description and output crs', async () => {
-    expect(client).not.toBeUndefined();
     const testPolygon =
       'POLYGON ((0.0 0.0, 0.0 80.0, 80.0 80.0, 80.0 0.0, 0.0 0.0))';
     const filter: WellFilter = {
@@ -275,7 +299,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - gets wells in geoJson polygon', async () => {
-    expect(client).not.toBeUndefined();
     const testPolygon = <GeoJson>{
       type: 'Polygon',
       coordinates: [
@@ -302,7 +325,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - get all wells with edm source', async () => {
-    expect(client).not.toBeUndefined();
     const testPolygon =
       'POLYGON ((0.0 0.0, 0.0 80.0, 80.0 80.0, 80.0 0.0, 0.0 0.0))';
     const filter: WellFilter = {
@@ -317,7 +339,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - get all wells with trajectory', async () => {
-    expect(client).not.toBeUndefined();
     const filter: WellFilter = { hasTrajectory: {} };
     const wells = await client.wells.filter(filter);
 
@@ -330,7 +351,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - get all wells with trajectory in range', async () => {
-    expect(client).not.toBeUndefined();
     const filter: WellFilter = {
       hasTrajectory: { minDepth: 1.0, maxDepth: 1.0 },
     };
@@ -345,7 +365,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - get all wells with trajectory - no returned in range', async () => {
-    expect(client).not.toBeUndefined();
     const filter: WellFilter = { hasTrajectory: { minDepth: 0, maxDepth: 0 } };
     const wells = await client.wells.filter(filter);
 
@@ -355,15 +374,13 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - get all block labels', async () => {
-    expect(client).not.toBeUndefined();
     const blocks: String[] = await client.wells.blocks();
 
-    expect(blocks).toContain('A');
+    expect(blocks).toContain('34/10');
     expect(blocks).toContain('B');
   });
 
   test('filter - get all field labels', async () => {
-    expect(client).not.toBeUndefined();
     const fields = await client.wells.fields();
 
     expect(fields).toContain('A');
@@ -371,30 +388,25 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - get all operator labels', async () => {
-    expect(client).not.toBeUndefined();
     const operators = await client.wells.operators();
 
-    expect(operators).toContain('A');
     expect(operators).toContain('B');
+    expect(operators).toContain('Op1');
   });
 
   test('filter - get all quadrants labels', async () => {
-    expect(client).not.toBeUndefined();
     const quadrants = await client.wells.quadrants();
 
     expect(quadrants).toContain('8');
-    expect(quadrants).toContain('B');
   });
 
   test('filter - get all source labels', async () => {
-    expect(client).not.toBeUndefined();
     const sources = await client.wells.sources();
 
     expect(sources).toContain('EDM');
   });
 
   test('filter - get all measurement types', async () => {
-    expect(client).not.toBeUndefined();
     const quadrants = await client.wells.measurements();
 
     expect(quadrants).not.toBeUndefined();
@@ -403,8 +415,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - has gamma measurement', async () => {
-    expect(client).not.toBeUndefined();
-
     const measurementFilter: MeasurementFilter = {
       measurementType: MeasurementType.GammaRay,
     };
@@ -424,8 +434,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - has multiple measurements, must match on all', async () => {
-    expect(client).not.toBeUndefined();
-
     const gammaRayFilter: MeasurementFilter = {
       measurementType: MeasurementType.GammaRay,
     };
@@ -448,8 +456,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - has multiple measurements, fail to match on all', async () => {
-    expect(client).not.toBeUndefined();
-
     const gammaRayFilter: MeasurementFilter = {
       measurementType: MeasurementType.GammaRay,
     };
@@ -471,8 +477,6 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
   });
 
   test('filter - has multiple measurements, can match on any', async () => {
-    expect(client).not.toBeUndefined();
-
     const gammaRayFilter: MeasurementFilter = {
       measurementType: MeasurementType.GammaRay,
     };
@@ -559,7 +563,7 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
     };
     const filter: WellFilter = { spudDate: dateFilter };
 
-    const wells = await client.wells.filter(filter);
+    const wells = await client.wells.filter(filter, undefined, 1);
 
     expect(wells).not.toBeUndefined();
 
@@ -589,5 +593,66 @@ describeIfCondition('CogniteClient setup in wells - integration test', () => {
 
     const externalIds = wells.items.map(well => well.externalId);
     expect(externalIds).toContain('well:34/10-8');
+  });
+
+  test('filter well on npt depth and duration', async () => {
+    const filter: WellFilter = {
+      npt: {
+        duration: {max: 30.0},
+        measuredDepth: { max: 3000.0, unit: LengthUnitEnum.METER }
+      },
+    };
+
+    const wells = await client.wells.filter(filter);
+    const externalIds = wells.items.map(well => well.externalId);
+    expect(externalIds).toContain('well:34/10-8');
+  });
+
+  test('filter well on npt criterias matching on all', async () => {
+
+    const filter: WellFilter = {
+      npt: {
+      duration: { min: 1.0, max: 3000.0 },
+      measuredDepth: { min: 1800, max: 3000.0, unit: LengthUnitEnum.METER },
+      nptCodes: { containsAll: ["FJSB", "GSLB"]},
+      nptCodeDetails: { containsAll: ["SLSN", "OLSF"]},
+      },
+    };
+
+    const wells = await client.wells.filter(filter);
+    const externalIds = wells.items.map(well => well.externalId);
+    expect(externalIds).toContain('well:34/10-8');
+  });
+
+  test('filter well on npt criterias matching on any', async () => {
+
+    const filter: WellFilter = {
+      npt: {
+      duration: { min: 1.0, max: 3000.0 },
+      measuredDepth: { min: 1800, max: 3000.0, unit: LengthUnitEnum.METER },
+      nptCodes: { containsAny: ["FJSB", "GSLB", "XSLC"]},
+      nptCodeDetails: { containsAny: ["SLSN", "OLSF", "ZJST"]},
+      },
+    };
+
+    const wells = await client.wells.filter(filter);
+    const externalIds = wells.items.map(well => well.externalId);
+
+    const actual = new Set(externalIds);
+    const expected = new Set(['well:34/10-24', 'well:34/10-8'])
+    const intersection = [...actual].filter(x => expected.has(x!));
+
+    expect(intersection.length).toBe(expected.size);
+  });
+
+  test('filter well on npt depth', async () => {
+    const filter: WellFilter = {
+      npt: {
+        measuredDepth: { min: 500000.0, unit: LengthUnitEnum.METER }
+      },
+    };
+
+    const wells = await client.wells.filter(filter);
+    expect(wells).not.toBeUndefined();
   });
 });