+++ packages/wells/src/client/api/casingsApi.ts
@@ -15,7 +15,7 @@ export class CasingsAPI extends ConfigureAPI {
         )
       )
       .catch(err => {
-        throw new HttpError(err.status, err.errorMessage, {});
+        throw new HttpError(err.status, err.data.error.message, {});
       });
   };
 
@@ -31,22 +31,13 @@ export class CasingsAPI extends ConfigureAPI {
         )
       )
       .catch(err => {
-        throw new HttpError(err.status, err.errorMessage, {});
+        throw new HttpError(err.status, err.data.error.message, {});
       });
   };
 
   private addLazyMethodsForCasing = (casing: Sequence): Sequence => {
     return <Sequence>{
-      id: casing.id,
-      columns: casing.columns,
-      createdTime: casing.createdTime,
-      lastUpdatedTime: casing.lastUpdatedTime,
-      name: casing.name,
-      description: casing.description,
-      assetId: casing.assetId,
-      externalId: casing.externalId,
-      metadata: casing.metadata,
-      dataSetId: casing.dataSetId,
+      ...casing,
       data: async (
         start?: number,
         end?: number,
@@ -89,7 +80,7 @@ export class CasingsAPI extends ConfigureAPI {
       .asyncPost<SequenceData>(path, { data: request })
       .then(response => response.data)
       .catch(err => {
-        throw new HttpError(err.status, err.errorMessage, {});
+        throw new HttpError(err.status, err.data.error.message, {});
       });
   };
 }