+++ packages/wells/src/client/model/NPTFilter.ts
@@ -12,13 +12,18 @@ export interface NPTFilter {
    */
   duration?: DoubleRange;
   /**
-   * @type {string}
+   * @type {string[]}
    * @memberof NPTFilter
    */
-  nptCode?: string;
+  nptCodes?: string[];
   /**
-   * @type {string}
+   * @type {string[]}
    * @memberof NPTFilter
    */
-  nptCodeDetail?: string;
+  nptCodeDetails?: string[];
+  /**
+   * @type {number[]}
+   * @memberof NPTFilter
+   */
+  wellboreIds?: number[];
 }